refactor(utils): cluster utility functions together
