--- src/utils/generate-new-id.ts
@@ -1,3 +1,0 @@
-import { nanoid } from "nanoid";
-
-export const generateNewId = () => nanoid(8);
